Move all contribution graph files to src/features/
