+++ apps/app/src/features/contribution-graph/services/aggregation-service.spec.ts
