+++ apps/app/src/features/contribution-graph/utils/contribution-graph-utils.ts
